+++ src/lab/mod.rs
@@ -1,0 +1,27 @@
+// Copyright 2025 Gabriel Bjørnager Jensen.
+//
+// This Source Code Form is subject to the terms of
+// the Mozilla Public License, v. 2.0. If a copy of
+// the MPL was not distributed with this file, you
+// can obtain one at:
+// <https://mozilla.org/MPL/2.0/>.
+
+//! LAB- and LCh-based colour facilities.
+
+mod cie_lab;
+mod cie_laba;
+mod cie_lch;
+mod cie_lcha;
+mod ok_lab;
+mod ok_laba;
+mod ok_lch;
+mod ok_lcha;
+
+pub use cie_lab::CieLab;
+pub use cie_laba::CieLaba;
+pub use cie_lch::CieLch;
+pub use cie_lcha::CieLcha;
+pub use ok_lab::OkLab;
+pub use ok_laba::OkLaba;
+pub use ok_lch::OkLch;
+pub use ok_lcha::OkLcha;