+++ src/rgb/op_rgba/mod.rs
@@ -7,23 +7,23 @@
 // <https://mozilla.org/MPL/2.0/>.
 
 use crate::{Colour, Component, DefinedGamut};
-use crate::rgb::{Rgba, SRgb};
+use crate::rgb::{OpRgb, Rgba};
 
-/// An sRGBA colour.
+/// An opRGBA colour.
 ///
 /// This type is guaranteed to always have the exact same layout as <code>[Rgba]&lt;T&gt;</code>, although with a different representation.
 /// Namely, channel order is preserved in memory.
 ///
-/// This type is equivalent to [`SRgb`] with an added alpha component.
+/// This type is equivalent to [`OpRgb`] with an added alpha component.
 #[repr(transparent)]
 #[derive(Clone, Copy, Debug, Default, Eq, Ord, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
 #[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
 #[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
-pub struct SRgba<T>(Rgba<T>);
+pub struct OpRgba<T>(Rgba<T>);
 
-impl<T: Component> SRgba<T> {
-	/// Constructs a new sRGBA colour.
+impl<T: Component> OpRgba<T> {
+	/// Constructs a new opRGBA colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn new(red: T, green: T, blue: T, alpha: T) -> Self {
@@ -31,22 +31,22 @@ impl<T: Component> SRgba<T> {
 		Self(colour)
 	}
 
-	/// Discards the sRGBA colour's alpha channel.
+	/// Discards the opRGBA colour's alpha channel.
 	#[inline(always)]
 	#[must_use]
-	pub const fn discard_alpha(self) -> SRgb<T> {
+	pub const fn discard_alpha(self) -> OpRgb<T> {
 		let (red, green, blue, _) = self.get();
-		SRgb::new(red, green, blue)
+		OpRgb::new(red, green, blue)
 	}
 
-	/// Reinterprets the sRGBA colour as a raw RGB colour.
+	/// Reinterprets the opRGBA colour as a raw RGB colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn as_rgba(self) -> Rgba<T> {
 		self.0
 	}
 
-	/// Deconstructs an sRGBA colour.
+	/// Deconstructs an RGBA colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn get(self) -> (T, T, T, T) {
@@ -54,6 +54,6 @@ impl<T: Component> SRgba<T> {
 	}
 }
 
-impl<T: Component> Colour for SRgba<T> { }
+impl<T: Component> Colour for OpRgba<T> { }
 
-impl<T: Component> DefinedGamut for SRgba<T> { }
+impl<T: Component> DefinedGamut for OpRgba<T> { }